Count peak-feature overlap with increasing distance

Count peak-feature overlap with increasing distance

peak_feature_ranged_counter(
  peaksL = NULL,
  featuresL = NULL,
  randomsL = NULL,
  category_colname = "hit",
  default_category = "hit",
  expand_kb = NULL,
  expand_range_kb = c(5, 5000),
  expand_factor = 2,
  do_random = FALSE,
  verbose = FALSE,
  ...
)

Arguments

peaksL
list named by experiment, containing a list of GRanges, where each set of GRanges represents a set of peaks.

featuresL
list named by experiment, containing GRanges representing features, with category_colname containing category values.

randomsL
list named by experiment, containing GRanges from which random peaks should be drawn, using the same number of peaks as each corresponding entry in peaksL.

category_colname
character string indicating the colname present in the GRanges for each featuresL element. If the values are factors, the factor levels should be preserved in order. When the column category_colname does not exist, or for any NA entries, default_category is used as a replacement.

default_category
vector length=1 used when category_colname is not present in colnames(values(featuresL[[i]])).

expand_kb
numeric vector of distances in kilobases (kb) to use when expanding the peaks in peaksL. When expand_kb is NULL then the default values are used, within the range defined by expand_range_kb.

expand_range_kb
numeric range of distances to allow in kilobases (kb) for expand_kb.

expand_factor
numeric value multiplied by the expand width during GenomicRanges::resize(), since the resize happens at peak center and defines a fixed width. For example, use expand_factor=2 when this distance should represent distance from the peak in either direction, for which the peaks are expanded from center by this distance.

do_random
logical indicating whether to choose random peaks from randomsL using the same number of peaks as the corresponding entry in peaksL.

verbose
logical indicating whether to print verbose output.

...
additional arguments are ignored.

Details

This function counts features overlaps across peaks, while progressively expanding the peak width. It also tabulates counts by feature category, using the feature annotations in category_colname.