Test hypergeometric enrichment of range count categories

Test hypergeometric enrichment of range count categories

peak_feature_ranged_enrichment(
  range_counts,
  test_cats = NULL,
  return_type = c("phyper", "wide"),
  verbose = FALSE,
  ...
)

Arguments

range_counts
data.frame output from peak_feature_ranged_counter().

test_cats
list of character vectors, where each vector contains one or more category values to test for enrichment against all other category values. When test_cats is NULL then each category value is tested individually against the other category values.

return_type
character string indicating the type of output to return: "phyper" will return the results of the hypergeometric test using phyper(); and "wide" will return a wide data.frame that contains the exact numbers used for this test.

verbose
logical indicating whether to print verbose output.

...
additional arguments are ignored.

Details

This function takes output from peak_feature_ranged_counter() and runs hypergeometric tests for each corresponding category value.

Specifically, the argument test_cats defines the category values to test in each enrichment test. For example, the category values may include c("-1", "0", "1"), and one may want to test c("1"). This test will count the number of peaks with category "1" that overlap features, compared to the number of peaks with other category values, in this case c("-1", "0"), that overlap features.

It is possible to combine category values, for example one may test c("-1", "1") and in this scenario, any peak with the category "1" or "-1" will be tested against all other peaks, in this case category "0".