Plot ranged enrichment hypergeometric results

Plot ranged enrichment hypergeometric results

plot_ranged_enrichment(
  paramsDF,
  p_cutoff = 0.01,
  p_color = "grey40",
  draw_points = TRUE,
  lwd = 1,
  pt_cex = 3,
  color_sub = NULL,
  fill_sub = color_sub,
  facet_formula = Experiment ~ PeakSet,
  shape_sub = c(higher = 24, lower = 25),
  alpha_sub = c(Test = 1, Random = 0.05),
  y_include = c(0, 10),
  x_include = c(5),
  test_cat_title = "Category",
  PeakSet_title = "PeakSet",
  main = "Enrichment by distance from peaks",
  xlab = "Distance from peak center (kb)",
  ylab = "-log10(hypergeometric P-value)",
  scales = "free_y",
  ...
)

Arguments

paramsDF
data.frame produced by peak_feature_ranged_enrichment() with return_type="phyper".

p_cutoff
numeric threshold used to draw a horizontal line indicating a threshold for statistical significance.

p_color
character with R color, used to color the horizontal line defined by p_cutoff.

draw_points
logical indicating whether each point should be drawn, or when draw_points=FALSE only the line is drawn.

lwd
numeric indicating the line width.

pt_cex
numeric indicating point size, used when draw_points=TRUE.

color_sub
character vector of R colors, whose names correspond to category values tested for enrichment.

fill_sub
character vector, same as color_sub but used when a fill color is required.

facet_formula
formula sufficient to inform facet_grid() used to create separate ggplot panels.

shape_sub
integer vector of R point shapes, whose names are peak names tested for enrichment.

alpha_sub
numeric vector of alpha transparency values ranging between 0 and 1, whose names are Test and Random. These values are only used when random permutation tests have been performed, and the entries that resulted from random permutations are stored in column "Test" and have value "Random".

y_include
numeric vector used to extend the y-axis range to include at least these values. This effect is useful in requiring the y-axis includes y=0 for example, and y=10 so that the y-axis range spans at least a reasonable viewing range for interpreting the data. The y-axis units are -log10(pvalue) values.

x_include
numeric vector used to extend the x-axis to ensure the range includes this value.

test_cat_title
character string used as the title for the category legend.

PeakSet_title
character string used as a legend for point shape, only relevant when draw_points=TRUE.

main
character string used as the title, called with ggtitle().

xlab, ylab
character string used for the x-axis and y-axis label, respectively.

scales
character string passed to facet_grid() which defines the y-axis and x-axis range across multiple facet panels: "free_y" will allow each panel to have independent y-axis range; "fixed" will use one y-axis range across all facet panels.

...
additional arguments are ignored.

Value

ggplot object, which can be viewed by calling print(), or further modified.

Details

This function takes "phyper" output from peak_feature_ranged_enrichment() and produces a plot using ggplot().